requirements.txt
torch>=2.1.0
transformers>=4.44.0
accelerate>=0.33.0
peft>=0.10.0
datasets>=2.18.0
evaluate>=0.4.2
sentencepiece>=0.2.0
jiwer>=3.0.0

opencv-python-headless>=4.10.0.0
Pillow>=10.3.0
numpy>=1.26.0
pandas>=2.2.0
albumentations>=1.4.7
tqdm>=4.66.0
pdf2image>=1.17.0

tinydb>=4.8.0

Flask>=3.0.0
